Update Vec3 to use an array internally
This is to make index based lookups more efficient

+++ src/bvh_node.rs
@@ -58,7 +58,7 @@ impl BvhNode {
         BvhNode { left, right, bbox }
     }
 
-    fn box_compare(a: Arc<dyn Hittable>, b: Arc<dyn Hittable>, axis_index: i32) -> Ordering {
+    fn box_compare(a: Arc<dyn Hittable>, b: Arc<dyn Hittable>, axis_index: usize) -> Ordering {
         let a_axis_interval = a.bounding_box().axis_interval(axis_index);
         let b_axis_interval = b.bounding_box().axis_interval(axis_index);
